+++ src/features/search/ui/components/EmptySearchResult.ts
