+++ src/shared/ui/disabledMoreButton.ts
@@ -1,0 +1,6 @@
+export function disableMoreButton(totalPages: number, currentPage: number) {
+  const $moreMoviesButton = document.getElementById("more-movies-button");
+  if (totalPages === currentPage) {
+    $moreMoviesButton?.classList.add("disabled");
+  }
+}
