+++ src/shared/ui/showSkeletons.ts
@@ -1,0 +1,5 @@
+import { createSkeletons } from "./createSkeletons";
+
+export function showSkeletons($container: HTMLElement, count: number = 10) {
+  $container.appendChild(createSkeletons(count));
+}
